+++ components/remote-controls/api-backend/src/commonMain/kotlin/com/flipperdevices/ifrmvp/backend/model/DeviceCategory.kt
@@ -10,4 +10,6 @@ class DeviceCategory(
     val id: Long,
     @SerialName("meta")
     val meta: CategoryMeta,
+    @SerialName("folder_name")
+    val folderName: String
 )