+++ components/remote-controls/api-backend/src/commonMain/kotlin/com/flipperdevices/ifrmvp/backend/model/IfrFileModel.kt
@@ -9,4 +9,8 @@ data class IfrFileModel(
     val id: Long,
     @SerialName("brand_id")
     val brandId: Long,
+    @SerialName("file_name")
+    val fileName: String,
+    @SerialName("folder_name")
+    val folderName: String
 )