+++ components/remote-controls/api-backend/src/commonMain/kotlin/com/flipperdevices/ifrmvp/backend/model/InfraredsResponse.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.ifrmvp.backend.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class InfraredsResponse(
+    @SerialName("infrared_files")
+    val infraredFiles: List<IfrFileModel>
+)