+++ components/remote-controls/brands/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/brands/presentation/decompose/internal/BrandsDecomposeComponentImpl.kt
@@ -21,11 +21,13 @@ import kotlinx.coroutines.flow.stateIn
 import me.gulya.anvil.assisted.ContributesAssistedFactory
 import javax.inject.Provider
 
+@Suppress("LongParameterList")
 @ContributesAssistedFactory(AppGraph::class, BrandsDecomposeComponent.Factory::class)
 class BrandsDecomposeComponentImpl @AssistedInject constructor(
     @Assisted componentContext: ComponentContext,
     @Assisted private val onBackClick: DecomposeOnBackParameter,
     @Assisted private val onBrandClick: (brandId: Long, brandName: String) -> Unit,
+    @Assisted private val onBrandLongClick: (brandId: Long) -> Unit,
     @Assisted categoryId: Long,
     createBrandsListViewModel: BrandsListViewModel.Factory,
     createQueryViewModel: Provider<QueryViewModel>
@@ -83,6 +85,10 @@ class BrandsDecomposeComponentImpl @AssistedInject constructor(
 
     override fun onBrandClick(brandModel: BrandModel) {
         onBrandClick.invoke(brandModel.id, brandModel.name)
+    }
+
+    override fun onBrandLongClick(brandModel: BrandModel) {
+        onBrandLongClick.invoke(brandModel.id)
     }
 
     override fun tryLoad() {