+++ components/remote-controls/brands/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/brands/presentation/decompose/internal/BrandsScreenDecomposeComponentImpl.kt
@@ -17,13 +17,15 @@ class BrandsScreenDecomposeComponentImpl @AssistedInject constructor(
     @Assisted categoryId: Long,
     @Assisted onBackClick: () -> Unit,
     @Assisted onBrandClick: (brandId: Long, brandName: String) -> Unit,
+    @Assisted onBrandLongClick: (brandId: Long) -> Unit,
     brandsDecomposeComponentFactory: BrandsDecomposeComponent.Factory,
 ) : BrandsScreenDecomposeComponent(componentContext) {
     private val brandsComponent = brandsDecomposeComponentFactory.createBrandsComponent(
         componentContext = childContext("BrandsComponent"),
         categoryId = categoryId,
         onBackClick = onBackClick,
-        onBrandClick = onBrandClick
+        onBrandClick = onBrandClick,
+        onBrandLongClick = onBrandLongClick
     )
 
     @Composable