+++ components/remote-controls/categories/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/categories/composable/components/DeviceCategoriesLoadingContent.kt
@@ -36,6 +36,7 @@ fun DeviceCategoriesLoadingContent(modifier: Modifier = Modifier) {
                         onClick = {},
                         deviceCategory = DeviceCategory(
                             id = -1,
+                            folderName = "",
                             meta = CategoryMeta(
                                 iconPngBase64 = "",
                                 iconSvgBase64 = "",