+++ components/remote-controls/core-model/src/commonMain/kotlin/com/flipperdevices/ifrmvp/model/serialization/ButtonDataEncoder.kt
@@ -5,6 +5,9 @@ import com.flipperdevices.ifrmvp.model.buttondata.ButtonData
 import com.flipperdevices.ifrmvp.model.buttondata.ChannelButtonData
 import com.flipperdevices.ifrmvp.model.buttondata.IconButtonData
 import com.flipperdevices.ifrmvp.model.buttondata.NavigationButtonData
+import com.flipperdevices.ifrmvp.model.buttondata.OkNavigationButtonData
+import com.flipperdevices.ifrmvp.model.buttondata.PowerButtonData
+import com.flipperdevices.ifrmvp.model.buttondata.ShutterButtonData
 import com.flipperdevices.ifrmvp.model.buttondata.TextButtonData
 import com.flipperdevices.ifrmvp.model.buttondata.UnknownButtonData
 import com.flipperdevices.ifrmvp.model.buttondata.VolumeButtonData
@@ -13,7 +16,7 @@ import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.encodeToJsonElement
 import kotlinx.serialization.json.jsonObject
 
-class ButtonDataEncoder(private val json: Json) {
+internal class ButtonDataEncoder(private val json: Json) {
     fun encodeToJsonObject(buttonData: ButtonData): JsonObject {
         return when (buttonData.type) {
             ButtonData.ButtonType.UNKNOWN -> {
@@ -36,8 +39,8 @@ class ButtonDataEncoder(private val json: Json) {
                 json.encodeToJsonElement(buttonData)
             }
 
-            ButtonData.ButtonType.NAVIGATION -> {
-                buttonData as NavigationButtonData
+            ButtonData.ButtonType.OK_NAVIGATION -> {
+                buttonData as OkNavigationButtonData
                 json.encodeToJsonElement(buttonData)
             }
 
@@ -48,6 +51,21 @@ class ButtonDataEncoder(private val json: Json) {
 
             ButtonData.ButtonType.ICON -> {
                 buttonData as IconButtonData
+                json.encodeToJsonElement(buttonData)
+            }
+
+            ButtonData.ButtonType.SHUTTER -> {
+                buttonData as ShutterButtonData
+                json.encodeToJsonElement(buttonData)
+            }
+
+            ButtonData.ButtonType.NAVIGATION -> {
+                buttonData as NavigationButtonData
+                json.encodeToJsonElement(buttonData)
+            }
+
+            ButtonData.ButtonType.POWER -> {
+                buttonData as PowerButtonData
                 json.encodeToJsonElement(buttonData)
             }
         }.jsonObject