+++ components/remote-controls/core-ui/src/main/kotlin/com/flipperdevices/ifrmvp/core/ui/button/core/SquareIconButton.kt
@@ -4,6 +4,7 @@ import android.content.res.Configuration.UI_MODE_NIGHT_YES
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -13,7 +14,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.core.ui.theme.LocalPalletV2
 import com.flipperdevices.ifrmvp.core.ui.ext.asPainter
-import com.flipperdevices.ifrmvp.core.ui.ext.tintFor
 import com.flipperdevices.ifrmvp.core.ui.layout.core.sf
 import com.flipperdevices.ifrmvp.model.buttondata.IconButtonData
 
@@ -71,18 +71,20 @@ fun SquareImageButton(
 fun SquareIconButton(
     iconType: IconButtonData.IconType,
     modifier: Modifier = Modifier,
+    iconTint: Color = MaterialTheme.colors.onPrimary,
+    background: Color = LocalPalletV2.current.surface.menu.body.dufault,
     contentDescription: String? = null,
     onClick: () -> Unit,
 ) {
     SquareButton(
         modifier = modifier,
         onClick = onClick,
-        background = LocalPalletV2.current.surface.menu.body.dufault
+        background = background
     ) {
         Icon(
             painter = iconType.asPainter(),
             contentDescription = contentDescription,
-            tint = iconType.tintFor(),
+            tint = iconTint,
             modifier = Modifier
                 .fillMaxSize()
                 .padding(12.sf)