+++ components/remote-controls/grid/saved/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/local/composable/LocalGridComposable.kt
@@ -86,7 +86,7 @@ fun LocalGridComposable(
             LocalGridComposableContent(
                 localGridComponent = localGridComponent,
                 model = model,
-                modifier = Modifier.padding(scaffoldPaddings)
+                modifier = Modifier.padding(scaffoldPaddings).navigationBarsPadding()
             )
             Box(
                 modifier = Modifier